requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-openai
langchain-community
langgraph
langfuse
chromadb
pdfplumber
pydantic
sqlalchemy
python-multipart
